use size_t to indicate the buffer size in mnl_cb_run*()

+++ src/callback.c
@@ -66,7 +66,7 @@ static mnl_cb_t default_cb_array[NLMSG_MIN_TYPE] = {
  *
  * This function propagates the callback return value.
  */
-int mnl_cb_run2(const char *buf, int numbytes, unsigned int seq,
+int mnl_cb_run2(const char *buf, size_t numbytes, unsigned int seq,
 		unsigned int portid, mnl_cb_t cb_data, void *data,
 		mnl_cb_t *cb_ctl_array, unsigned int cb_ctl_array_len)
 {
@@ -128,7 +128,7 @@ out:
  *
  * This function propagates the callback return value.
  */
-int mnl_cb_run(const char *buf, int numbytes, unsigned int seq,
+int mnl_cb_run(const char *buf, size_t numbytes, unsigned int seq,
 	       unsigned int portid, mnl_cb_t cb_data, void *data)
 {
 	return mnl_cb_run2(buf, numbytes, seq, portid, cb_data, data, NULL, 0);
